LPS-47563 Extract base class to toplevel
