+++ portal-service/src/com/liferay/portal/kernel/search/SearchResultManagerUtil.java
@@ -16,7 +16,6 @@ package com.liferay.portal.kernel.search;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.util.PortalUtil;
 import com.liferay.portlet.documentlibrary.model.DLFileEntryConstants;
 import com.liferay.portlet.messageboards.model.MBMessage;
 import com.liferay.registry.ServiceReference;
@@ -90,33 +89,6 @@ public class SearchResultManagerUtil {
 				}
 
 			});
-
-	private static class BaseSearchResultManager
-		implements SearchResultManager {
-
-		@Override
-		public SearchResult createSearchResult(Document document)
-			throws PortalException {
-
-			long classNameId = GetterUtil.getLong(
-				document.get(Field.CLASS_NAME_ID));
-			long classPK = GetterUtil.getLong(document.get(Field.CLASS_PK));
-
-			if ((classPK > 0) && (classNameId > 0)) {
-				String className = PortalUtil.getClassName(classNameId);
-
-				return new SearchResult(className, classPK);
-			}
-
-			String entryClassName = GetterUtil.getString(
-				document.get(Field.ENTRY_CLASS_NAME));
-			long entryClassPK = GetterUtil.getLong(
-				document.get(Field.ENTRY_CLASS_PK));
-
-			return new SearchResult(entryClassName, entryClassPK);
-		}
-
-	}
 
 	private static class DefaultSearchResultManagerImpl
 		implements SearchResultManager {
